feat(Student): 학생 í핸들ã러 기능 구현 및 포스트 맨 확인 완료

+++ src/main/java/com/example/dormi/mapper/DormiMapperService.java
@@ -52,7 +52,7 @@ public class DormiMapperService {
     return mapper.deleteStudent(studentId, studentHakbun);
   }
 
-  public StudentInfoVo selectStudentOneByIdHakbun(long studentId, long studentHakbun) {
+  public StudentInfoVo selectStudentOneByIdHakbun(int studentId, long studentHakbun) {
     return mapper.selectStudentOneByIdHakbun(studentId, studentHakbun);
   }
